Updated annotations schema and made a special client-only collection.
A special client-only collection LocalAnnotations mirrors Annotations
collection, but allows adding temporary client-only annotations. We use
this to be able to add temporary annotations for a current user which
are not stored on the server until user inserts some real data. Then we
upgrade a client-only annotation to a real annotation.