Moved meta menu icons to the right.
Maybe it really looks better this way. Additionally, some trickery would
be needed to have meta menu be correctly opening to the left, because
setting overflow to auto crops all the content inside of the
.annotations. So meta menu cannot really grow to the left by the
default. (You have to move left edge of .annotations to the left to make
space and then move it back with padding.)