Refactored permissions so that they can be stacked.
This allows admins to automatically have both maintainer and read
access. And maintainers to automatically have read access. This means
that it is not necessary to keep access fields in special sync, but each
field can really be used to store information about users at that
particular permission and we combine fields together when making
queries.