Fixed switchToFallbackWorker code path.
It didn't really work. When it switched we went consuming queue before
we reset it to the fallback worker, so callbacks were still using web
worker. Similar error reporting didn't work because we first have to
destroy, then call callback. Now it is also possible to force using web
worker.