Improved when we rerender publications.
Previously, we rerendered a publication only when its display fields
changed, but ignored any other rerendered callback. In fact this is not
good enough because sometimes display wrapper DOM element gets changed
and we do have to rerender into it. This was happening regularly when
one switched from closed access to open access, in another browser
publication did not automatically render now (it did, but into wrong
display wrapper DOM element not in the DOM anymore).