Removed optimization which was breaking annotation filter.
We had an optimization which assumed that once bounding boxes of
highlights are computed they do not change. This is not true, sometimes
bounding boxes contain NaN and afterwards are filled with correct
values. So now variable of all highlights is a normal reactive variable
which triggers change also when values change, not only keys. But there
are more logic so that values do not change if not strictly necessary.
The issue was easy reproducible by having annotations which are linked
to a highlight, going from the publication view to some other view and
then pressing back in the browser. Those linked annotations did not
appear because their highlights contained NaN and reactive value was
never updated with the correct values.