Removed dependencies between jobs.
We are not really using them. We are mostly using that all jobs relate
to the same publication. Moreover, this now allows that spawned jobs
are processed immediately (if there are multiple worker instances) or
at least that if parent job fails at the end (or process is killed),
spawned jobs will still be done.