Removing Mendeley logo for now.
It signaled wrong message. We put it there because we were planing of
using their API to add metadata and allow users to import annotations
they made in Mendeley. Because their API mentions CC-BY license, to be
clear of any issues, we prepared to attribute them by placing a logo on
our about page. They said that would be enough to satisfy the
attribution requirement. In no other way Mendeley has any influence
over the project.