Fixed importing of PDFs in Safari. Fixes #627
Not sure why blob was added as a dependency anyway. It should be used
only when testing, to fix tests on PhantomJS. But now we are putting it
as a proper dependency of those packages which need it so things work
correctly when testing them. We are not yet using it in production,
probably those browsers have larger issues anyway. We could use it in
production now though, because new version of meteor-blob has now a
version which does not break Safari anymore.